Add WebRTC's FEC processing to ExternalOutput
WebRTC does FEC processing to reduce corruption in video due to packet loss.  Let's implement that!  I stripped out just the FEC portion of WebRTC and incorporated that.  One important note: WebRTC has its own endianness detection that's implemented in rtp_utility.cc; it is very important to get that right or you're find yourself exploding inside std::list in a really confusing way.  I think the routine as-is will work for licode across the board, but we probably want to double-check that.

Recording should be unaffected; the only consequence of this change should be "less corruption" in output video.  Corruption is still possible until ExternalOutput starts receiving retransmitted packets coming in on the Rtx SSRC channel.