Change WebRtcConnection to use webrtc's fec if necessary
because firefox doesn't support RED/FEC, licode was converting RED/FEC to raw VP8, but it was doing this for everybody, regardless of whether or not they could digest red.  change it so for endpoints that can digest red, we send it straight through.  For those that can't, we use WebRTC's FEC handling to recover missing packets.