Do not send FIR frames constantly if we're recording
This is both A) a violation of the WebRTC spec and B) explodes image quality and network bandwidth, particularly for multi-party chat where the coordinated FIR requests decimate network quality on a periodic basis.  At some point we should to a NACK or PLI in response this, but a NACK is absolutely the wrong thing to do.