fix sdp offer bundle parsing when streaming video-only or audio-only
The fix is to simply change the 4 to a 3 so that the conditional code to add the BundleTag objects to the bundleTags vector.  Credit to Jeremy Noring on the lynckia google group for finding this fix.

In the case of video-only `a=group:BUNDLE video` or audio-only `a=group:BUNDLE audio`, the number of parts is only 3.  Without the bundle tags, the remote description had an invalid bundle definition `a=group:BUNDLE`.  Most of the time with WebRTC the bundle definition will be `a=group:BUNDLE audio video` for audio and video streaming.

This fixes video-only and audio-only streaming in Chrome 47.